(#526) introduce (print) for REPL
