+++ src/ebisp/repl.c
@@ -5,42 +5,9 @@
 #include "interpreter.h"
 #include "scope.h"
 #include "gc.h"
+#include "repl_runtime.h"
 
 #define REPL_BUFFER_MAX 1024
-
-static struct EvalResult gc_inspect_adapter(void *param, Gc *gc, struct Scope *scope, struct Expr args)
-{
-    assert(gc);
-    assert(scope);
-    (void) param;
-    (void) args;
-
-    gc_inspect(gc);
-
-    return eval_success(NIL(gc));
-}
-
-static struct EvalResult quit(void *param, Gc *gc, struct Scope *scope, struct Expr args)
-{
-    assert(gc);
-    assert(scope);
-    (void) args;
-    (void) param;
-
-    exit(0);
-
-    return eval_success(NIL(gc));
-}
-
-static struct EvalResult get_scope(void *param, Gc *gc, struct Scope *scope, struct Expr args)
-{
-    assert(gc);
-    assert(scope);
-    (void) param;
-    (void) args;
-
-    return eval_success(scope->expr);
-}
 
 static void eval_line(Gc *gc, Scope *scope, const char *line)
 {
@@ -81,9 +48,7 @@ int main(int argc, char *argv[])
         .expr = CONS(gc, NIL(gc), NIL(gc))
     };
 
-    set_scope_value(gc, &scope, SYMBOL(gc, "quit"), NATIVE(gc, quit, NULL));
-    set_scope_value(gc, &scope, SYMBOL(gc, "gc-inspect"), NATIVE(gc, gc_inspect_adapter, NULL));
-    set_scope_value(gc, &scope, SYMBOL(gc, "scope"), NATIVE(gc, get_scope, NULL));
+    load_repl_runtime(gc, &scope);
 
     while (true) {
         printf("> ");
